backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6
langchain==0.1.5
langchain-openai==0.0.5
langgraph==0.0.20
supabase>=2.3.0,<3.0.0
httpx>=0.24.0,<0.25.0
openai==1.10.0
elevenlabs==0.2.27
pandas==2.1.4
numpy==1.26.3
pydantic==2.5.3
pytest==7.4.4
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pypdf==3.17.1
